+++ R/getMainCategory.R
@@ -2,14 +2,17 @@
 #'
 #' Get daum news main category names and ids recently.
 #'
+#' @param fresh If TRUE, get data from internet. 
+#'              Default is FALSE which is return with cache.
 #' @return Get data.frame(chr:cate_name, chr:url).
 #' @export
-#' @importFrom xml2 read_html
-#' @importFrom rvest html_nodes html_text html_attr
-
-getMainCategory <- function() {
+#' @importFrom rvest read_html html_nodes html_text html_attr
+getMainCategory <- function(fresh = FALSE) {
+  if (!fresh) {
+    return(category_main)
+  }
   root <- "http://media.daum.net/breakingnews"
-  hobj <- xml2::read_html(root)
+  hobj <- rvest::read_html(root)
   hobj_nodes <-
     rvest::html_nodes(hobj, "div#mArticle ul.tab_nav li a")
   titles <- rvest::html_text(hobj_nodes)
@@ -22,5 +25,4 @@ getMainCategory <- function() {
                stringsAsFactors = F)
   urls <- urls[-1, ]
   return(urls)
-  
 }